Introduce maven structure and drop support for most versions
