+++ .gitignore
@@ -1,2 +1,3 @@
 /*.iml
 /.idea/
+/target/
