+++ src/main/java/net/blackscarx/betterchairs/ChairType.java
@@ -1,0 +1,5 @@
+package net.blackscarx.betterchairs;
+
+public enum ChairType {
+	STAIR, SLAP;
+}
