+++ src/main/resources/plugin.yml
@@ -1,0 +1,38 @@
+name: ${pluginName}
+#description: 'Plugin description for ${pluginName}'
+version: ${version}
+api-version: '1.15'
+
+main: ${groupId}.${artifactId}.ChairsPlugin
+
+author: BlackScarx
+authors: ['BlackScarx', 'Sprax2013']
+website: 'https://github.com/Sprax2013/BetterChairs'
+
+commands:
+  betterchairsreload:
+    description: 'Reload the config'
+    aliases:
+      - 'bcreload'
+      - 'bcr'
+    permission: ${pluginName}.reload
+  betterchairstoggle:
+    description: 'Command betterchairstoggle'
+    aliases:
+      - 'bctoggle'
+      - 'bct'
+    permission: ${pluginName}.toggle
+  betterchairsreset:
+    description: 'Command betterchairsreset'
+    aliases:
+      - 'bcreset'
+    permission: ${pluginName}.reset
+
+permissions:
+  ${pluginName}.use:
+    description: Use the stairs
+  ${pluginName}.regen:
+    description: Regen when you are sit on chairs
+  ${pluginName}.update:
+    descritpion: Permission for betterchairs update
+    default: false
