--- src/net/blackscarx/betterchairs/ChairType.java
@@ -1,5 +1,0 @@
-package net.blackscarx.betterchairs;
-
-public enum ChairType {
-	STAIR, SLAP;
-}
